Add capability for conn to put itself back into the pool
Add capability for a connection to put itself back into the pool if:

*   explicitly told to
*   all existing references are removed/deleted and the connection is
    still valid (connected==True).

This is achieve by passing the current pool to the client upon
creation. The client can then optionally choose to put iself back into
the pool, if the connector class supports it.

This is likely a breaking change, due to the connector class
instantiation signature modification.

fix for issue #5