make distinct the max number of connections handled from the pool size.
Before that changes the maximum number of connections created was equal
to the max of connections you could have in a pool. So if a connection
was out more longer than expected in the pool, it wasn't possible to
emit more connections. This change make sure that we can do that.

Defaults are 10 connections max kept in the pool and 150 maximum
connections that can be emitted. To get the old behaviour you can use
max_conn = max_size.
Also while i'm here make sure the con