remove the max_connections option
We can't be sure that a user will always use the pool object to close a
connection and reduce the number of max connections (and it wasn't
really done if the connection was closed in restkit). It is better then
to just remove this option which was blocking most of the usecases
around.